ml_pipeline/requirements.txt
# MedScript ML Pipeline — Pinned Dependencies
# Python 3.10.x required

# Core
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.1
tqdm==4.66.1
scipy==1.11.4

# PyTorch — choose ONE:
# GPU:  pip install torch==2.0.1+cu118 torchvision==0.15.2+cu118 --index-url https://download.pytorch.org/whl/cu118
# CPU:  pip install torch==2.0.1+cpu torchvision==0.15.2+cpu --index-url https://download.pytorch.org/whl/cpu
torch==2.0.1
torchvision==0.15.2

# HuggingFace
transformers==4.35.2
datasets==2.14.6
tokenizers==0.14.1
accelerate==0.24.1
evaluate==0.4.1
timm==0.9.7
sentencepiece==0.1.99

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
albumentations==1.3.1

# NLP / Matching
rapidfuzz==3.5.2
python-Levenshtein==0.23.0
networkx==3.2.1

# Serving
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.2
